NUP-2397: Rename TP.py to BacktrackingTM.py
