+++ docs/examples/opf/model_params.py
@@ -123,7 +123,7 @@ MODEL_PARAMS = {
         'tmParams': {
             # TP diagnostic output verbosity control;
             # 0: silent; [1..6]: increasing levels of verbosity
-            # (see verbosity in nupic/trunk/py/nupic/research/TP.py and TP10X*.py)
+            # (see verbosity in nupic/trunk/py/nupic/research/BacktrackingTM.py and TP10X*.py)
             'verbosity': 0,
 
             # Number of cell columns in the cortical region (same number for
