+++ examples/opf/experiments/anomaly/spatial/10field_many_balanced/description.py
@@ -200,7 +200,7 @@ config = {
         'tmParams': {
             # TP diagnostic output verbosity control;
             # 0: silent; [1..6]: increasing levels of verbosity
-            # (see verbosity in nupic/trunk/py/nupic/research/TP.py and TP10X*.py)
+            # (see verbosity in nupic/trunk/py/nupic/research/BacktrackingTM.py and TP10X*.py)
             'verbosity': 0,
 
             # Number of cell columns in the cortical region (same number for
