+++ examples/opf/experiments/anomaly/temporal/hotgym/description.py
@@ -212,7 +212,7 @@ config = {
         'tmParams': {
             # TP diagnostic output verbosity control;
             # 0: silent; [1..6]: increasing levels of verbosity
-            # (see verbosity in nupic/trunk/py/nupic/research/TP.py and TP10X*.py)
+            # (see verbosity in nupic/trunk/py/nupic/research/BacktrackingTM.py and TP10X*.py)
             'verbosity': 0,
 
             # Number of cell columns in the cortical region (same number for
