+++ scripts/profiling/tp_profile.py
@@ -25,7 +25,7 @@ import sys
 import numpy
 # chose desired TP implementation to compare:
 from nupic.research.TP10X2 import TP10X2 as CppTP 
-from nupic.research.TP import TP as PyTP
+from nupic.research.BacktrackingTM import TP as PyTP
 
 
 def profileTP(tpClass, tpDim, nRuns):
