+++ src/nupic/research/TP10X2.py
@@ -23,7 +23,7 @@ import numpy
 from numpy import *
 
 import nupic.math
-from nupic.research.TP import TP
+from nupic.research.BacktrackingTM import TP
 
 from nupic.bindings.algorithms import Cells4
