+++ tests/integration/nupic/algorithms/tp_overlapping_sequences_test.py
@@ -46,7 +46,7 @@ import pprint
 import random
 import unittest2 as unittest
 
-from nupic.research.TP import TP
+from nupic.research.BacktrackingTM import TP
 from nupic.research.TP10X2 import TP10X2
 from nupic.research import fdrutilities as fdrutils
 from nupic.support.unittesthelpers import testcasebase
