+++ tests/unit/nupic/research/tp_constant_test.py
@@ -29,7 +29,7 @@ import numpy as np
 import unittest2 as unittest
 
 from nupic.research import fdrutilities as fdrutils
-from nupic.research.TP import TP
+from nupic.research.BacktrackingTM import TP
 from nupic.research.TP10X2 import TP10X2
 
 _SEED = 42
@@ -59,7 +59,7 @@ def _getSimplePatterns(numOnes, numPatterns):
 
 
 def _createTps(numCols):
-  """Create two instances of temporal poolers (TP.py and TP10X2.py) with
+  """Create two instances of temporal poolers (BacktrackingTM.py and TP10X2.py) with
   identical parameter settings."""
 
   # Keep these fixed:
